ImageService updated

+++ Model/Services/ImageService/Resources/Output/ImageOutput.cs
@@ -9,7 +9,7 @@ namespace Es.Udc.DotNet.PracticaMaD.Model.Services.ImageService.Resources.Output
     [Serializable]
     public class ImageOutput
     {
-        public ImageOutput(byte[] imageData, string title, int userId, int categoryId, int likes)
+        public ImageOutput(byte[] imageData, string title, long userId, long categoryId, int likes)
         {
             ImageData = imageData;
             Title = title;
@@ -20,8 +20,8 @@ namespace Es.Udc.DotNet.PracticaMaD.Model.Services.ImageService.Resources.Output
 
         public byte[] ImageData { get; private set; }
         public string Title { get; private set; }
-        public int CategoryId { get; private set; }
-        public int UserId { get; private set; }
+        public long CategoryId { get; private set; }
+        public long UserId { get; private set; }
         public int Likes { get; private set; }
     }
 }
